// frontend/src/types/auth.ts
export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  email: string;
  firstName: string;
  lastName: string;
  role: string;
}

export interface AuthUser {
  email: string;
  firstName: string;
  lastName: string;
  role: string;
}